Refactor referral processing and risk scoring logic
Introduces ReferralProcessResult to clarify referral outcomes and refactors referral processing to improve error handling and separation of validation and scoring. Adds lookback_days to RiskScoreConfig and ensures it is loaded from config. Fixes type for ip_abuse_score in risk scoring input.

+++ crates/gem_rewards/src/risk_scoring/client.rs
@@ -29,7 +29,7 @@ impl RiskScoringInput {
             ip_country_code: self.ip_result.country_code.clone(),
             ip_usage_type: self.ip_result.usage_type.clone(),
             ip_isp: self.ip_result.isp.clone(),
-            ip_abuse_score: self.ip_result.confidence_score as i64,
+            ip_abuse_score: self.ip_result.confidence_score,
             referrer_verified: self.referrer_verified,
         }
     }
